+++ src/main/java/de/marhali/easyi18n/io/parser/yaml/YamlParserStrategy.java
@@ -2,6 +2,7 @@ package de.marhali.easyi18n.io.parser.yaml;
 
 import com.intellij.openapi.vfs.VirtualFile;
 
+import de.marhali.easyi18n.exception.SyntaxException;
 import de.marhali.easyi18n.io.parser.ParserStrategy;
 import de.marhali.easyi18n.model.TranslationData;
 import de.marhali.easyi18n.model.TranslationFile;
@@ -10,6 +11,7 @@ import de.marhali.easyi18n.settings.ProjectSettings;
 
 import org.jetbrains.annotations.NotNull;
 
+import org.yaml.snakeyaml.error.YAMLException;
 import thito.nodeflow.config.MapSection;
 import thito.nodeflow.config.Section;
 
@@ -34,7 +36,14 @@ public class YamlParserStrategy extends ParserStrategy {
         TranslationNode targetNode = super.getOrCreateTargetNode(file, data);
 
         try(Reader reader = new InputStreamReader(vf.getInputStream(), vf.getCharset())) {
-            Section input = Section.parseToMap(reader);
+            Section input;
+
+            try {
+                input = Section.parseToMap(reader);
+            } catch (YAMLException ex) {
+                throw new SyntaxException(ex.getMessage(), file);
+            }
+
             YamlMapper.read(file.getLocale(), input, targetNode);
         }
     }